+++ src/Umbraco.Tests/PublishedContent/DynamicNodeTests.cs
@@ -3,9 +3,11 @@ using System.IO;
 using NUnit.Framework;
 using Umbraco.Core;
 using Umbraco.Core.Configuration;
+using Umbraco.Core.IO;
 using Umbraco.Tests.TestHelpers;
 using Umbraco.Web;
-using umbraco.IO;
+using Umbraco.Web.PublishedCache;
+using Umbraco.Web.PublishedCache.XmlPublishedCache;
 using umbraco.MacroEngines;
 using umbraco.NodeFactory;
 using System.Linq;
@@ -15,9 +17,12 @@ namespace Umbraco.Tests.PublishedContent
     [TestFixture]
     public class DynamicNodeTests : DynamicDocumentTestsBase<DynamicNode, DynamicNodeList>
     {
-        protected override bool RequiresDbSetup
+        /// <summary>
+        /// We only need a new schema per fixture... speeds up testing
+        /// </summary>
+        protected override DatabaseBehavior DatabaseTestBehavior
         {
-            get { return true; }
+            get { return DatabaseBehavior.NewSchemaPerFixture; }
         }
 
         public override void Initialize()
@@ -25,21 +30,22 @@ namespace Umbraco.Tests.PublishedContent
             base.Initialize();
             //copy the umbraco settings file over
             var currDir = new DirectoryInfo(TestHelper.CurrentAssemblyDirectory);
-            File.Copy(
-                currDir.Parent.Parent.Parent.GetDirectories("Umbraco.Web.UI")
-                    .First()
-                    .GetDirectories("config").First()
-                    .GetFiles("umbracoSettings.Release.config").First().FullName,
-                Path.Combine(currDir.Parent.Parent.FullName, "config", "umbracoSettings.config"),
-                true);
+
+            var configPath = Path.Combine(currDir.Parent.Parent.FullName, "config");
+            if (Directory.Exists(configPath) == false)
+                Directory.CreateDirectory(configPath);
+
+            var umbracoSettingsFile = Path.Combine(currDir.Parent.Parent.FullName, "config", "umbracoSettings.config");
+            if (File.Exists(umbracoSettingsFile) == false)
+                File.Copy(
+                    currDir.Parent.Parent.Parent.GetDirectories("Umbraco.Web.UI")
+                        .First()
+                        .GetDirectories("config").First()
+                        .GetFiles("umbracoSettings.Release.config").First().FullName,
+                    Path.Combine(currDir.Parent.Parent.FullName, "config", "umbracoSettings.config"),
+                    true);
 
             UmbracoSettings.SettingsFilePath = IOHelper.MapPath(SystemDirectories.Config + Path.DirectorySeparatorChar, false);
-
-            //for testing, we'll specify which assemblies are scanned for the PluginTypeResolver
-            PluginManager.Current.AssembliesToScan = new[]
-				{
-					typeof(DynamicNode).Assembly
-				};
 
             //need to specify a custom callback for unit tests
             DynamicNode.GetDataTypeCallback = (docTypeAlias, propertyAlias) =>
@@ -47,7 +53,7 @@ namespace Umbraco.Tests.PublishedContent
                 if (propertyAlias == "content")
                 {
                     //return the rte type id
-                    return Guid.Parse("5e9b75ae-face-41c8-b47e-5f4b0fd82f83");
+                    return Guid.Parse(Constants.PropertyEditors.TinyMCEv3);
                 }
                 return Guid.Empty;
             };
@@ -77,11 +83,15 @@ namespace Umbraco.Tests.PublishedContent
 
         public override void TearDown()
         {
+            //TODO: Deleting the umbracoSettings.config file makes a lot of tests fail
+
+            //var currDir = new DirectoryInfo(TestHelper.CurrentAssemblyDirectory);
+
+            //var umbracoSettingsFile = Path.Combine(currDir.Parent.Parent.FullName, "config", "umbracoSettings.config");
+            //if (File.Exists(umbracoSettingsFile))
+            //    File.Delete(umbracoSettingsFile);
+
             base.TearDown();
-
-            PluginManager.Current.AssembliesToScan = null;
-
-            UmbracoSettings.ResetSetters();
         }
 
         protected override dynamic GetDynamicNode(int id)
@@ -89,10 +99,13 @@ namespace Umbraco.Tests.PublishedContent
             //var template = Template.MakeNew("test", new User(0));
             //var ctx = GetUmbracoContext("/test", template.Id);
             var ctx = GetUmbracoContext("/test", 1234);
-            var contentStore = new DefaultPublishedContentStore();
+
+            var cache = ctx.ContentCache.InnerCache as PublishedContentCache;
+            if (cache == null) throw new Exception("Unsupported IPublishedContentCache, only the Xml one is supported.");
+
             var node = new DynamicNode(
                 new DynamicBackingItem(
-                    new Node(ctx.GetXml().SelectSingleNode("//*[@id='" + id + "' and @isDoc]"))));
+                    new Node(cache.GetXml(ctx, ctx.InPreviewMode).SelectSingleNode("//*[@id='" + id + "' and @isDoc]"))));
             Assert.IsNotNull(node);
             return (dynamic)node;
         }
