+++ src/Umbraco.Tests/Routing/NiceUrlProviderTests.cs
@@ -5,6 +5,7 @@ using System.IO;
 using System.Linq;
 using NUnit.Framework;
 using Umbraco.Tests.TestHelpers;
+using Umbraco.Web.PublishedCache.XmlPublishedCache;
 using Umbraco.Web.Routing;
 
 namespace Umbraco.Tests.Routing
@@ -17,27 +18,47 @@ namespace Umbraco.Tests.Routing
 			base.Initialize();
 
             var currDir = new DirectoryInfo(TestHelper.CurrentAssemblyDirectory);
-            File.Copy(
-                currDir.Parent.Parent.Parent.GetDirectories("Umbraco.Web.UI")
-                    .First()
-                    .GetDirectories("config").First()
-                    .GetFiles("umbracoSettings.Release.config").First().FullName,
-                Path.Combine(currDir.Parent.Parent.FullName, "config", "umbracoSettings.config"),
-                true);
+
+            var configPath = Path.Combine(currDir.Parent.Parent.FullName, "config");
+            if (Directory.Exists(configPath) == false)
+                Directory.CreateDirectory(configPath);
+
+            var umbracoSettingsFile = Path.Combine(currDir.Parent.Parent.FullName, "config", "umbracoSettings.config");
+            if (File.Exists(umbracoSettingsFile) == false)
+                File.Copy(
+                    currDir.Parent.Parent.Parent.GetDirectories("Umbraco.Web.UI")
+                        .First()
+                        .GetDirectories("config").First()
+                        .GetFiles("umbracoSettings.Release.config").First().FullName,
+                    Path.Combine(currDir.Parent.Parent.FullName, "config", "umbracoSettings.config"),
+                    true);
 
             SettingsForTests.SettingsFilePath = Core.IO.IOHelper.MapPath(Core.IO.SystemDirectories.Config + Path.DirectorySeparatorChar, false);
 		}
 
-		internal override IRoutesCache GetRoutesCache()
-		{
-			return new DefaultRoutesCache(false);
-		}
+
+        public void TearDown()
+        {
+            //TODO: Deleting the umbracoSettings.config file makes a lot of tests fail
+
+            //var currDir = new DirectoryInfo(TestHelper.CurrentAssemblyDirectory);
+
+            //var umbracoSettingsFile = Path.Combine(currDir.Parent.Parent.FullName, "config", "umbracoSettings.config");
+            //if (File.Exists(umbracoSettingsFile))
+            //    File.Delete(umbracoSettingsFile);
+        }
+
+        protected override void FreezeResolution()
+        {
+            SiteDomainHelperResolver.Current = new SiteDomainHelperResolver(new SiteDomainHelper());
+
+            base.FreezeResolution();
+        }
 
 		/// <summary>
 		/// This checks that when we retreive a NiceUrl for multiple items that there are no issues with cache overlap 
 		/// and that they are all cached correctly.
         /// </summary>
-        [Ignore]
 		[Test]
 		public void Ensure_Cache_Is_Correct()
 		{
@@ -59,18 +80,20 @@ namespace Umbraco.Tests.Routing
 
 			foreach (var sample in samples)
 			{
-				var result = routingContext.NiceUrlProvider.GetNiceUrl(sample.Key);
+				var result = routingContext.UrlProvider.GetUrl(sample.Key);
 				Assert.AreEqual(sample.Value, result);
 			}
 
 			var randomSample = new KeyValuePair<int, string>(1177, "/home/sub1/custom-sub-1");
 			for (int i = 0; i < 5; i++)
 			{
-				var result = routingContext.NiceUrlProvider.GetNiceUrl(randomSample.Key);
+				var result = routingContext.UrlProvider.GetUrl(randomSample.Key);
 				Assert.AreEqual(randomSample.Value, result);
 			}
 
-			var cachedRoutes = ((DefaultRoutesCache)routingContext.UmbracoContext.RoutesCache).GetCachedRoutes();
+            var cache = routingContext.UmbracoContext.ContentCache.InnerCache as PublishedContentCache;
+            if (cache == null) throw new Exception("Unsupported IPublishedContentCache, only the Xml one is supported.");
+            var cachedRoutes = cache.RoutesCache.GetCachedRoutes();
 			Assert.AreEqual(8, cachedRoutes.Count);
 
 			foreach (var sample in samples)
@@ -79,7 +102,7 @@ namespace Umbraco.Tests.Routing
 				Assert.AreEqual(sample.Value, cachedRoutes[sample.Key]);
 			}
 
-			var cachedIds = ((DefaultRoutesCache)routingContext.UmbracoContext.RoutesCache).GetCachedIds();
+			var cachedIds = cache.RoutesCache.GetCachedIds();
 			Assert.AreEqual(8, cachedIds.Count);
 
 			foreach (var sample in samples)
@@ -107,7 +130,7 @@ namespace Umbraco.Tests.Routing
 		    SettingsForTests.HideTopLevelNodeFromPath = false;
 		    SettingsForTests.UseDomainPrefixes = false;
 
-			var result = routingContext.NiceUrlProvider.GetNiceUrl(nodeId);
+			var result = routingContext.UrlProvider.GetUrl(nodeId);
 			Assert.AreEqual(niceUrlMatch, result);
 		}
 
@@ -130,7 +153,7 @@ namespace Umbraco.Tests.Routing
             SettingsForTests.HideTopLevelNodeFromPath = true;
             SettingsForTests.UseDomainPrefixes = false;
 
-			var result = routingContext.NiceUrlProvider.GetNiceUrl(nodeId);
+			var result = routingContext.UrlProvider.GetUrl(nodeId);
 			Assert.AreEqual(niceUrlMatch, result);
 		}
 
@@ -143,14 +166,14 @@ namespace Umbraco.Tests.Routing
             SettingsForTests.HideTopLevelNodeFromPath = false;
 
             SettingsForTests.UseDomainPrefixes = false;
-			Assert.AreEqual("/home/sub1/custom-sub-1/", routingContext.NiceUrlProvider.GetNiceUrl(1177));
+			Assert.AreEqual("/home/sub1/custom-sub-1/", routingContext.UrlProvider.GetUrl(1177));
 
             SettingsForTests.UseDomainPrefixes = true;
-			Assert.AreEqual("http://example.com/home/sub1/custom-sub-1/", routingContext.NiceUrlProvider.GetNiceUrl(1177));
+			Assert.AreEqual("http://example.com/home/sub1/custom-sub-1/", routingContext.UrlProvider.GetUrl(1177));
 
             SettingsForTests.UseDomainPrefixes = false;
-			routingContext.NiceUrlProvider.EnforceAbsoluteUrls = true;
-			Assert.AreEqual("http://example.com/home/sub1/custom-sub-1/", routingContext.NiceUrlProvider.GetNiceUrl(1177));
+			routingContext.UrlProvider.Mode = UrlProviderMode.Absolute;
+			Assert.AreEqual("http://example.com/home/sub1/custom-sub-1/", routingContext.UrlProvider.GetUrl(1177));
 		}
 
 		[Test]
@@ -162,12 +185,12 @@ namespace Umbraco.Tests.Routing
             SettingsForTests.HideTopLevelNodeFromPath = false;
 
             SettingsForTests.UseDomainPrefixes = false;
-			Assert.AreEqual("#", routingContext.NiceUrlProvider.GetNiceUrl(999999));
+			Assert.AreEqual("#", routingContext.UrlProvider.GetUrl(999999));
             SettingsForTests.UseDomainPrefixes = true;
-			Assert.AreEqual("#", routingContext.NiceUrlProvider.GetNiceUrl(999999));
+			Assert.AreEqual("#", routingContext.UrlProvider.GetUrl(999999));
             SettingsForTests.UseDomainPrefixes = false;
-			routingContext.NiceUrlProvider.EnforceAbsoluteUrls = true;
-			Assert.AreEqual("#", routingContext.NiceUrlProvider.GetNiceUrl(999999));
+			routingContext.UrlProvider.Mode = UrlProviderMode.Absolute;
+			Assert.AreEqual("#", routingContext.UrlProvider.GetUrl(999999));
 		}
 	}
 }
